logic for changing wpa_supplicant.conf file
Add shell script, cp_wpa_conf.sh, which moves a conf file from some directory into the /etc/wpa_supplicant directory so that the new file will be used when the system reboots.
Add testUSBmounted.py which tests the logic for detecting a USB drive, and if a wpa_supplicant.conf file is present, uses the cp_wpa_conf.sh script to move that file into /etc/wpa_supplicant.  This short program tests the concept and functionality of the process.
Update cordiebot2.py to use the logic for moving the wpa_supplicant.conf file to /etc/wpa_supplicant.  cordiebot2.py now looks for a USB drive with the name CORDIEBOT, and if it is found and has a wpa_supplicant.conf file on it, moves that file to /etc/wpa_supplicant, and advised that the system will restart (which is required for the new network and password to be used) and that the USB drive should be removed.
The cordiebot2.py program also has sections added with headers for definitions, classes, and various operations to better organize the program visually.